requirements.txt
opencv-python
Pillow
pytesseract
requests
numpy
pygrabber
deep-translator
langdetect
